Geocoder Service API Parameter Reference

This dataset contains the mapping that allows this package to use a universal syntax to specify parameters for different geocoder services.

The "generic_name" field is a universal field name while the api_name field shows the specific parameter name for the given geocoder service (method). When the api_name is missing this means that the parameter is not supported by the given geocoder service. When generic_name is missing this means the parameter is specific to that geocoding service.

Reference the documentation for geo for more information. Also reference vignette("tidygeocoder") for more details on constructing API queries.

api_parameter_reference

Format

A tibble dataframe

method

Geocoder service name

generic_name

Universal parameter name

api_name

Name of the parameter for the specified geocoder service

default_value

Default value of the parameter

required

Is the parameter required by the specified geocoder service?

Source

Links to API documentation for each geocoder service are below.

Census

Nominatim ("osm")

Geocodio

Location IQ ("iq")

See also