Extract reverse geocoding results

Parses the output of the query_api function for reverse geoocding. The address is extracted into the first column of the returned dataframe. This function is not used for batch geocoded results. Refer to query_api for example usage.

extract_reverse_results(method, response, full_results = TRUE, flatten = TRUE)

Arguments

method
method name

response
content from the geocoder service (returned by the query_api function)

full_results
if TRUE then the full results (not just an address column) will be returned.

flatten
if TRUE then flatten any nested dataframe content

Value

geocoder results in tibble format

See also