Geocode addresses

First attempts to use the US Census Geocoder (geo_census) method and then uses the Nominatim/OSM (geo_osm) method if the census method failed. Returns latitude and longitude coordinates and the method used to return results (OSM or Census)

geo_cascade(address, lat = lat, long = long, verbose = FALSE)

Arguments

address
single line address.

lat
name of latitude field

long
name of longitude field

verbose
logical. If TRUE outputs logs.

Value

latitude and longitude coordinates and the geocoder method used (geo_method) in tibble format (3 columns)

Examples

# \donttest{
geo_cascade("1600 Pennsylvania Ave Washington, DC")

#> # A tibble: 1 x 3
#>     lat  long geo_method
#>   <dbl> <dbl> <chr>     
#> 1  38.9 -77.0 census    

geo_cascade("Paris, France")

#> # A tibble: 1 x 3
#>     lat  long geo_method
#>   <dbl> <dbl> <chr>     
#> 1  48.9  2.35 osm       

# }