Geocode addresses

Obtains latitude and longitude coordinates from an address using the Nominatim (OSM) geocoder service. Can be used with non-US or non-street level addresses unlike the Census geocoder. This function calls the geocode_OSM function from the tmaptools package.

geo_osm(address, lat = lat, long = long, verbose = FALSE)

Arguments

address
single line address

lat
name of latitude field

long
name of longitude field

verbose
logical. If TRUE outputs logs.

Value

latitude and longitude coordinates in tibble format

Details

WARNING - This service has a usage limit and it will return missing coordinates once the usage limit is reached.

Examples

# \donttest{
geo_osm("1600 Pennsylvania Ave Washington, DC")

#> # A tibble: 1 x 2
#>     lat  long
#>   <dbl> <dbl>
#> 1  38.9 -77.0

geo_osm("Paris, France",verbose=TRUE)

#> Paris, France

#> # A tibble: 1 x 2
#>     lat  long
#>   <dbl> <dbl>
#> 1  48.9  2.35

# }