Combine multiple geocoding queries

Executes multiple geocoding queries on a dataframe input and combines the results. To use a character vector input instead, see the geo_combine function.

geocode_combine(
  .tbl,
  queries,
  global_params = list(),
  query_names = NULL,
  return_list = FALSE,
  cascade = TRUE,
  lat = lat,
  long = long
)

Arguments

.tbl
dataframe containing addresses

queries
list of lists parameter. Each list contains parameters for a query (ie. list(list(method = 'osm'), list(method = 'census'), ...))

global_params
list parameter. Contains parameters to be used for all queries. (ie. list(full_results = TRUE, unique_only = TRUE))

query_names
an optional vector of names to use for labeling the queries.

return_list
if TRUE then results will be returned in a named list. If FALSE (default) then all results will be combined into a single dataframe.

cascade
if TRUE then only addresses that are not found will be attempted by the following query. If FALSE then all queries will attempt to geocode all addresses.

lat
latitude column name. Can be quoted or unquoted (ie. lat or 'lat').

long
longitude column name. Can be quoted or unquoted (ie. long or 'long').

Value

tibble (dataframe)

See also

Examples

# \donttest{
geocode_combine(sample_addresses,
 queries = list(list(method = 'census'), list(method = 'osm')),
 global_params = list(address = 'addr'), cascade = TRUE)

#> 

#> 

#> Passing 9 addresses to the US Census batch geocoder

#> Query completed in: 1.4 seconds

#> Passing 5 addresses to the Nominatim single address geocoder

#> Query completed in: 5 seconds

#> # A tibble: 9 × 5
#>   name                 addr                                     lat   long query
#>   <chr>                <chr>                                  <dbl>  <dbl> <chr>
#> 1 White House          1600 Pennsylvania Ave NW Washington, … 38.9   -77.0 cens…
#> 2 Transamerica Pyramid 600 Montgomery St, San Francisco, CA … 37.8  -122.  cens…
#> 3 NY Stock Exchange    11 Wall Street, New York, New York     40.7   -74.0 cens…
#> 4 Willis Tower         233 S Wacker Dr, Chicago, IL 60606     41.9   -87.6 cens…
#> 5 Chateau Frontenac    1 Rue des Carrieres, Quebec, QC G1R 4… 46.8   -71.2 osm  
#> 6 Nashville            Nashville, TN                          36.2   -86.8 osm  
#> 7 Nairobi              Nairobi, Kenya                         -1.30   36.8 osm  
#> 8 Istanbul             Istanbul, Turkey                       41.0    29.0 osm  
#> 9 Tokyo                Tokyo, Japan                           35.7   140.  osm  

# }