Geocoder service API parameter reference

This dataset contains the mapping that allows this package to use a universal syntax to specify parameters for different geocoder services. Note that latitude nad longitude input parameters for reverse geocoding are not in this dataset and are instead handled directly by the reverse_geo function.

The generic_name column is a universal parameter name that is shared between services. The api_name column is the parameter name for the given geocoder service specified by the method column. When generic_name is missing this means the parameter is specific to that geocoder service.

Reference the documentation for geo for more information. Also reference vignette("tidygeocoder") for more details on constructing API queries.

api_parameter_reference

Format

A tibble dataframe

method

Geocoder service name

generic_name

Universal parameter name

api_name

Name of the parameter for the specified geocoder service

default_value

Default value of the parameter

required

Is the parameter required by the specified geocoder service?

Source

Links to API documentation for each geocoder service are below.

Census

Nominatim ("osm")

Geocodio

Location IQ ("iq")

Google

OpenCage

Mapbox

HERE

See also