Geocoding service API parameter reference

This dataset contains the mapping that allows this package to use a universal syntax to specify parameters for different geocoding services. Note that latitude and longitude input parameters for reverse geocoding are not in this dataset and are instead handled directly by the reverse_geo function.

The generic_name column is a universal parameter name that is shared between services. The api_name column is the parameter name for the given geocoding service specified by the method column. When generic_name is missing this means the parameter is specific to that geocoding service.

While the "census" and "google" services do not have a limit argument in their APIs, tidygeocoder provides a passthrough so you can still use the limit argument in geo and reverse_geo to limit the number of results per input.

Note that some geocoding services only use the limit argument for forward geocoding. Refer to API documentation of each service for more information.

Reference the documentation for geo and reverse_geo for more information. Also reference vignette("tidygeocoder") for more details on constructing API queries.

api_parameter_reference

Format

A tibble dataframe

method

Geocoding service name

generic_name

Universal parameter name

api_name

Name of the parameter for the specified geocoding service

default_value

Default value of the parameter

required

Is the parameter required by the specified geocoding service?

Details

The API documentation for each service is linked to below:

Nominatim

US Census

ArcGIS

Geocodio

Location IQ

Google

OpenCage

Mapbox

HERE

TomTom

MapQuest

Bing

Geoapify

See also