Geocode addresses

Geocodes addresses given as character values. The geocode function utilizes this function on addresses contained in dataframes. See example usage in vignette("tidygeocoder").

Note that not all geocoder services support certain address component parameters. For example, the Census geocoder only covers the United States and does not have a "country" parameter.

Refer to api_parameter_reference, min_time_reference, and batch_limit_reference for more details on geocoder service parameters and usage.

This function uses the get_api_query, query_api, and extract_results functions to create, execute, and parse geocoder API queries.

geo(
  address = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  postalcode = NULL,
  country = NULL,
  method = "osm",
  cascade_order = c("census", "osm"),
  lat = lat,
  long = long,
  limit = 1,
  min_time = NULL,
  api_url = NULL,
  timeout = 20,
  mode = "",
  full_results = FALSE,
  unique_only = FALSE,
  return_addresses = TRUE,
  flatten = TRUE,
  batch_limit = NULL,
  batch_limit_error = TRUE,
  verbose = FALSE,
  no_query = FALSE,
  custom_query = list(),
  return_type = "locations",
  iq_region = "us",
  geocodio_v = 1.6,
  param_error = TRUE,
  mapbox_permanent = FALSE,
  here_request_id = NULL,
  mapquest_open = FALSE
)

Arguments

address
single line address (ie. '1600 Pennsylvania Ave NW, Washington, DC'). Do not combine with the address component arguments below (street, city, county, state, postalcode, country).

street
street address (ie. '1600 Pennsylvania Ave NW')

city
city (ie. 'Tokyo')

county
county (ie. 'Jefferson')

state
state (ie. 'Kentucky')

postalcode
postalcode (zip code if in the United States)

country
country (ie. 'Japan')

method
the geocoder service to be used. API keys are loaded from environmental variables. Run usethis::edit_r_environ() to open your .Renviron file and add an API key as an environmental variable. For example, add the line GEOCODIO_API_KEY="YourAPIKeyHere".

"osm": Nominatim.

"census": US Census. Geographic coverage is limited to the United States. Batch geocoding is supported.

"arcgis": ArcGIS.

"geocodio": Geocodio. Geographic coverage is limited to the United States and Canada. An API key must be stored in "GEOCODIO_API_KEY". Batch geocoding is supported.

"iq": Location IQ. An API key must be stored in "LOCATIONIQ_API_KEY".

"google": Google. An API key must be stored in "GOOGLEGEOCODE_API_KEY".

"opencage": OpenCage. An API key must be stored in "OPENCAGE_KEY".

"mapbox": Mapbox. An API key must be stored in "MAPBOX_API_KEY".

"here": HERE. An API key must be stored in "HERE_API_KEY". Batch geocoding is supported, but must be explicitly called with mode = "batch".

"tomtom": TomTom. An API key must be stored in "TOMTOM_API_KEY". Batch geocoding is supported.

"mapquest": MapQuest. An API key must be stored in "MAPQUEST_API_KEY". Batch geocoding is supported.

"bing": Bing. An API key must be stored in "BINGMAPS_API_KEY". Batch geocoding is supported, but must be explicitly called with mode = "batch".

"geoapify": Geoapify. An API key must be stored in "GEOAPIFY_KEY".

"cascade" : First uses one geocoder service and then uses a second geocoder service if the first service didn't return results. The services and order is specified by the cascade_order argument. Note that this is not compatible with full_results = TRUE as geocoder services have different columns that they return.

cascade_order
a vector with two character values for the method argument in the order in which the geocoder services will be attempted for method = "cascade" (ie. c("census", "geocodio"))

lat
latitude column name. Can be quoted or unquoted (ie. lat or "lat").

long
longitude column name. Can be quoted or unquoted (ie. long or "long").

limit
maximum number of results to return per input address. For many geocoder services the maximum value of the limit parameter is 100. Pass limit = NULL to use the default limit value of the selected geocoder service. For batch geocoding, limit must be set to 1 (default) if return_addresses = TRUE. Refer to api_parameter_reference for more details.

min_time
minimum amount of time for a query to take (in seconds). If NULL then min_time will be set to the default value specified in min_time_reference.

api_url
custom API URL. If specified, the default API URL will be overridden. This parameter can be used to specify a local Nominatim server, for instance.

timeout
query timeout (in minutes)

mode
set to 'batch' to force batch geocoding or 'single' to force single address geocoding (one address per query). If not specified then batch geocoding will be used if available (given method selected) when multiple addresses are provided; otherwise single address geocoding will be used. For the "here" and "bing" methods the batch mode should be explicitly specified with mode = 'batch'.

full_results
returns all data from the geocoder service if TRUE. If FALSE then only longitude and latitude are returned from the geocoder service.

unique_only
only return results for unique inputs if TRUE

return_addresses
return input addresses with results if TRUE. Note that most services return the input addresses with full_results = TRUE and setting return_addresses to FALSE does not prevent this.

flatten
if TRUE then any nested dataframes in results are flattened if possible. Note that in some cases results are flattened regardless such as for Geocodio batch geocoding.

batch_limit
limit to the number of addresses in a batch geocoding query. Defaults to the value in batch_limit_reference if not specified.

batch_limit_error
if TRUE then an error is thrown if the number of addresses exceeds the batch limit. (if executing a batch query). This is reverted to FALSE when using the cascade method.

verbose
if TRUE then detailed logs are output to the console

no_query
if TRUE then no queries are sent to the geocoder service and verbose is set to TRUE. Used for testing.

custom_query
API-specific parameters to be used, passed as a named list (ie. list(extratags = 1).

return_type
only used when method = "census". Two possible values:

"locations" (default)

"geographies": returns additional geography columns. See the Census geocoder API documentation for more details.

iq_region
"us" (default) or "eu". Used for establishing the API URL for the "iq" method.

geocodio_v
version of geocodio API. Used for establishing the API URL for the "geocodio" method.

param_error
if TRUE then an error will be thrown if any address parameters are used that are invalid for the selected service (method). If method = "cascade" then no errors will be thrown.

mapbox_permanent
if TRUE then the mapbox.places-permanent endpoint would be used. Note that this option should be used only if you have applied for a permanent account. Unsuccessful requests made by an account that does not have access to the endpoint may be billable.

here_request_id
This parameter would return a previous HERE batch job, identified by its RequestID. The RequestID of a batch job is displayed when verbose is TRUE. Note that this option would ignore the current address parameter on the request, so return_addresses needs to be FALSE.

mapquest_open
if TRUE then MapQuest would use the Open Geocoding endpoint, that relies solely on data contributed to OpenStreetMap.

Value

tibble (dataframe)

See also

Examples

# \donttest{
geo(street = "600 Peachtree Street NE", city = "Atlanta",
 state = "Georgia", method = "census")

#> # A tibble: 1 x 5
#>   street                  city    state     lat  long
#>   <chr>                   <chr>   <chr>   <dbl> <dbl>
#> 1 600 Peachtree Street NE Atlanta Georgia  33.8 -84.4

geo(address = c("Tokyo, Japan", "Lima, Peru", "Nairobi, Kenya"),
 method = 'osm')

#> # A tibble: 3 x 3
#>   address           lat  long
#>   <chr>           <dbl> <dbl>
#> 1 Tokyo, Japan    35.7  140. 
#> 2 Lima, Peru     -12.1  -77.0
#> 3 Nairobi, Kenya  -1.30  36.8

geo(county = 'Jefferson', state = "Kentucky", country = "US",
     method = 'osm')

#> # A tibble: 1 x 5
#>   county    state    country   lat  long
#>   <chr>     <chr>    <chr>   <dbl> <dbl>
#> 1 Jefferson Kentucky US       38.2 -85.7

# }