Geocode street addresses in a dataframe

Takes a dataframe containing addresses as a input. Returns the dataframe with latitude and longitude coordinate columns using a user specified geocoder function.

geocode(.tbl, address, method = "census", lat = lat, long = long, ...)

Arguments

.tbl
dataframe

address
name of column containing addresses in .tbl

method
the geocoder function you want to use

"census": geo_census - can only handle US street level addresses

"osm": geo_osm - more versatile than Census but has a usage limit

"cascade": geo_cascade - first tries to use census then tries osm

lat
name of latitude field

long
name of longitude field

...
arguments supplied to the relevant geocoder function

Value

input dataframe (.tbl) with latitude and longitude fields appended

Details

See example usage in vignette("tidygeocoder")

Examples

# \donttest{
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

sample_addresses %>% geocode(addr)

#> # A tibble: 8 x 4
#>   name                   addr                                         lat   long
#>   <chr>                  <chr>                                      <dbl>  <dbl>
#> 1 "White House"          "1600 Pennsylvania Ave Washington, DC"      38.9  -77.0
#> 2 "Transamerica Pyramid" "600 Montgomery St, San Francisco, CA 941…  37.8 -122. 
#> 3  NA                    "Fake Address"                              NA     NA  
#> 4  NA                     NA                                         NA     NA  
#> 5 ""                     ""                                          NA     NA  
#> 6 "US City"              "Nashville,TN"                              NA     NA  
#> 7 "Willis Tower"         "233 S Wacker Dr, Chicago, IL 60606"        41.9  -87.6
#> 8 "International City"   "Nairobi, Kenya"                            NA     NA  

sample_addresses %>% geocode(addr,method='cascade',lat=latitude,long=longitude)

#> # A tibble: 8 x 5
#>   name              addr                           latitude longitude geo_method
#>   <chr>             <chr>                             <dbl>     <dbl> <chr>     
#> 1 "White House"     "1600 Pennsylvania Ave Washin…    38.9      -77.0 census    
#> 2 "Transamerica Py… "600 Montgomery St, San Franc…    37.8     -122.  census    
#> 3  NA               "Fake Address"                    NA         NA   NA        
#> 4  NA                NA                               NA         NA   NA        
#> 5 ""                ""                                NA         NA   NA        
#> 6 "US City"         "Nashville,TN"                    36.2      -86.8 osm       
#> 7 "Willis Tower"    "233 S Wacker Dr, Chicago, IL…    41.9      -87.6 census    
#> 8 "International C… "Nairobi, Kenya"                  -1.28      36.8 osm       

# }