Query a Geocoder API

The get_api_query function can create queries for this function to execute.

query_api(
  api_url,
  query_parameters,
  mode = "single",
  batch_file = NULL,
  address_list = NULL,
  content_encoding = "UTF-8",
  timeout = 20
)

Arguments

api_url
Base URL of the API. query parameters are appended to this

query_parameters
api query parameters in the form of a named list

mode
"single" : geocode a single address (all methods) "list" : batch geocode a list of addresses (geocodio) "file" : batch geocode a file of addresses (census)

batch_file
a csv file of addresses to upload (census)

address_list
a list of addresses for batch geocoding (geocodio) should be 'json' for geocodio and 'multipart' for census

content_encoding
Encoding to be used for parsing content. Census uses "ISO-8859-1", all other services use "UTF-8"

timeout
timeout in minutes for batch geocoding

Value

raw results from the query