Execute a geocoder API query

The get_api_query function can create queries for this function to execute.

query_api(
  api_url,
  query_parameters,
  mode = "single",
  batch_file = NULL,
  input_list = NULL,
  content_encoding = "UTF-8",
  timeout = 20,
  method = ""
)

Arguments

api_url
Base URL of the API. query parameters are appended to this

query_parameters
api query parameters in the form of a named list

mode

"single" : geocode a single address (all methods)

"list" : batch geocode a list of addresses (geocodio)

"file" : batch geocode a file of addresses (census)

batch_file
a csv file of addresses to upload (census)

input_list
a list of addresses or latitude, longitude coordinates for batch geocoding (geocodio) should be 'json' for geocodio and 'multipart' for census

content_encoding
Encoding to be used for parsing content

timeout
timeout in minutes

method
if 'mapquest' then the query status code is changed appropriately

Value

a named list containing the response content (content) and the HTTP request status (status)

See also

Examples

# \donttest{
raw1 <- query_api("http://nominatim.openstreetmap.org/search", 
   get_api_query("osm", list(address = 'Hanoi, Vietnam')))
   
raw1$status

#> [1] 200

   
extract_results('osm', jsonlite::fromJSON(raw1$content))

#>        lat      lon  place_id
#> 1 21.02945 105.8544 309312568
#>                                                                  licence
#> 1 Data © OpenStreetMap contributors, ODbL 1.0. https://osm.org/copyright
#>   osm_type  osm_id                                      boundingbox
#> 1 relation 1903516 20.5645154, 21.3852777, 105.2848986, 106.0200725
#>       display_name    class           type importance
#> 1 Hà Nội, Việt Nam boundary administrative  0.6282962
#>                                                                                    icon
#> 1 https://nominatim.openstreetmap.org/ui/mapicons//poi_boundary_administrative.p.20.png

raw2 <- query_api("http://nominatim.openstreetmap.org/reverse", 
   get_api_query("osm", custom_parameters = list(lat = 38.895865, lon = -77.0307713)))
   
extract_reverse_results('osm', jsonlite::fromJSON(raw2$content))

#> # A tibble: 1 x 15
#>   display_name place_id licence osm_type osm_id osm_lat osm_lon tourism road 
#>   <chr>           <int> <chr>   <chr>     <int> <chr>   <chr>   <chr>   <chr>
#> 1 L’Enfant's …   3.02e8 Data ©… way      9.00e8 38.895… -77.03… L’Enfa… Penn…
#> # … with 6 more variables: city <chr>, state <chr>, postcode <chr>,
#> #   country <chr>, country_code <chr>, boundingbox <list>

# }