Reverse geocode coordinates

Reverse geocodes geographic coordinates (latitude and longitude) given as numeric values. Latitude and longitude inputs are limited to possible values. Latitudes must be between -90 and 90 and longitudes must be between -180 and 180. Invalid values will not be sent to the geocoder service. The reverse_geocode function utilizes this function on coordinates contained in dataframes. See example usage in vignette("tidygeocoder").

This function uses the get_api_query, query_api, and extract_reverse_results functions to create, execute, and parse the geocoder API queries.

reverse_geo(
  lat,
  long,
  method = "osm",
  address = address,
  limit = 1,
  min_time = NULL,
  api_url = NULL,
  timeout = 20,
  mode = "",
  full_results = FALSE,
  unique_only = FALSE,
  return_coords = TRUE,
  flatten = TRUE,
  batch_limit = NULL,
  verbose = FALSE,
  no_query = FALSE,
  custom_query = list(),
  iq_region = "us",
  geocodio_v = 1.6,
  mapbox_permanent = FALSE,
  here_request_id = NULL,
  mapquest_open = FALSE
)

Arguments

lat
latitude values (input data)

long
longitude values (input data)

method
the geocoder service to be used. Refer to api_parameter_reference and the API documentation for each geocoder service for usage details and limitations. Note that the Census service does not support reverse geocoding. Run usethis::edit_r_environ() to open your .Renviron file for editing to add API keys as an environmental variables.

"osm": Nominatim (OSM). Worldwide coverage.

"geocodio": Commercial geocoder. Covers US and Canada and has batch geocoding capabilities. Requires an API Key to be stored in the "GEOCODIO_API_KEY" environmental variable.

"iq": Commercial Nominatim geocoder service. Requires an API Key to be stored in the "LOCATIONIQ_API_KEY" environmental variable.

"google": Commercial Google geocoder service. Requires an API Key to be stored in the "GOOGLEGEOCODE_API_KEY" environmental variable.

"opencage": Commercial geocoder with various open data sources (e.g. OpenStreetMap). Requires an API Key to be stored in the "OPENCAGE_KEY" environmental variable.

"mapbox": Commercial Mapbox geocoder service. Requires an API Key to be stored in the "MAPBOX_API_KEY" environmental variable.

"here": Commercial HERE geocoder service. Requires an API Key to be stored in the "HERE_API_KEY" environmental variable. Can perform batch geocoding, but this must be specified with mode = 'batch'.

"tomtom": Commercial TomTom geocoder service. Requires an API Key to be stored in the "TOMTOM_API_KEY" environmental variable. Can perform batch geocoding.

"mapquest": Commercial MapQuest geocoder service. Requires an API Key to be stored in the "MAPQUEST_API_KEY" environmental variable. Can perform batch geocoding.

"bing": Commercial Bing geocoder service. Requires an API Key to be stored in the "BINGMAPS_API_KEY" environmental variable. Can perform batch geocoding.

address
name of the address column (output data)

limit
maximum number of results to return per coordinate For many geocoder services the maximum value for the limit parameter is 100. Use limit = NULL to use the default value of the selected geocoder service. For batch geocoding, limit must be set to 1 (default) if return_coords = TRUE.

min_time
minimum amount of time for a query to take (in seconds). If NULL then min_time will be set to the lowest value that complies with the usage requirements of the free tier of the selected geocoder service.

api_url
custom API URL. If specified, the default API URL will be overridden. This parameter can be used to specify a local Nominatim server.

timeout
query timeout (in minutes)

mode
set to 'batch' to force batch geocoding or 'single' to force single address geocoding (one coordinate per query). If not specified then batch geocoding will be used if available (given method selected) when multiple addresses are provided; otherwise single address geocoding will be used. For 'here' and 'bing' the batch mode should be explicitly enforced.

full_results
returns all data from the geocoder service if TRUE. If FALSE then only a single address column will be returned from the geocoder service.

unique_only
only return results for unique addresses if TRUE

return_coords
return input coordinates with results if TRUE. Note that most services return the input coordinates with full_results = TRUE and setting return_addresses to FALSE does not prevent this.

flatten
if TRUE then any nested dataframes in results are flattened if possible. Note that Geocodio batch geocoding results are flattened regardless.

batch_limit
limit to the number of addresses in a batch geocoding query. Both geocodio and census batch geocoders have a 10,000 limit so this is the default. 'here' has a 1,000,000 address limit. 'mapquest' has a 100 address limit. 'bing' as a 50 address limit.

verbose
if TRUE then detailed logs are output to the console

no_query
if TRUE then no queries are sent to the geocoder and verbose is set to TRUE

custom_query
API-specific parameters to be used, passed as a named list (ie. list(extratags = 1)).

iq_region
'us' (default) or 'eu'. Used for establishing API URL for the 'iq' method

geocodio_v
version of geocodio api. Used for establishing API URL for the 'geocodio' method.

mapbox_permanent
if TRUE then the mapbox.places-permanent endpoint would be used. Note that this option should be used only if you have applied for a permanent account. Unsuccessful requests made by an account that does not have access to the endpoint may be billable.

here_request_id
This parameter would return a previous HERE batch job, identified by its RequestID. The RequestID of a batch job is displayed when verbose is TRUE. Note that this option would ignore the current lat, long parameters on the request, so return_coords needs to be FALSE.

mapquest_open
if TRUE then MapQuest would use the Open Geocoding endpoint, that relies solely on data contributed to OpenStreetMap.

Value

parsed geocoding results in tibble format

See also

Examples

# \donttest{
 reverse_geo(lat = 38.895865, long = -77.0307713, method = 'osm', verbose = TRUE)

#> Number of Unique Coordinates: 1

#> Querying API URL: https://nominatim.openstreetmap.org/reverse

#> Passing the following parameters to the API:

#> limit : "1"

#> lat : "38.895865"

#> lon : "-77.0307713"

#> format : "json"

#> 

#> HTTP Status Code: 200

#> Query completed in: 0.1 seconds

#> 

#> # A tibble: 1 x 3
#>     lat  long address                                                           
#>   <dbl> <dbl> <chr>                                                             
#> 1  38.9 -77.0 L’Enfant's plan, Pennsylvania Avenue, Washington, District of Col…

 
 reverse_geo(lat = c(38.895865, 43.6534817, 300), 
 long = c(-77.0307713, -79.3839347, 600), method = 'osm', full_results = TRUE, verbose = TRUE)

#> Number of Unique Coordinates: 2

#> Executing single coordinate geocoding...

#> Number of Unique Coordinates: 1

#> Querying API URL: https://nominatim.openstreetmap.org/reverse

#> Passing the following parameters to the API:

#> limit : "1"

#> lat : "38.895865"

#> lon : "-77.0307713"

#> format : "json"

#> 

#> HTTP Status Code: 200

#> Query completed in: 0.5 seconds

#> 

#> Number of Unique Coordinates: 1

#> Querying API URL: https://nominatim.openstreetmap.org/reverse

#> Passing the following parameters to the API:

#> limit : "1"

#> lat : "43.6534817"

#> lon : "-79.3839347"

#> format : "json"

#> 

#> HTTP Status Code: 200

#> Query completed in: 0.5 seconds

#> 

#> # A tibble: 3 x 22
#>     lat  long address place_id licence osm_type  osm_id osm_lat osm_lon tourism
#>   <dbl> <dbl> <chr>      <int> <chr>   <chr>      <int> <chr>   <chr>   <chr>  
#> 1  38.9 -77.0 L’Enfa…   3.02e8 Data ©… way       9.00e8 38.895… -77.03… L’Enfa…
#> 2  43.7 -79.4 Toront…   1.37e8 Data ©… way       1.99e8 43.653… -79.38… NA     
#> 3  NA    NA   NA       NA      NA      NA       NA      NA      NA      NA     
#> # … with 12 more variables: road <chr>, city <chr>, state <chr>,
#> #   postcode <chr>, country <chr>, country_code <chr>, boundingbox <list>,
#> #   amenity <chr>, house_number <chr>, neighbourhood <chr>, quarter <chr>,
#> #   state_district <chr>

 
# }