Reverse geocode coordinates in a dataframe

Takes a dataframe containing coordinates (latitude and longitude) and returns the reverse geocoding query results from a specified service by using the reverse_geo function. See example usage in vignette("tidygeocoder").

This function passes all additional parameters (...) to the reverse_geo function, so you can refer to its documentation for more details on possible arguments.

reverse_geocode(
  .tbl,
  lat,
  long,
  address = address,
  return_input = TRUE,
  limit = 1,
  return_coords = NULL,
  unique_only = FALSE,
  ...
)

Arguments

.tbl
dataframe containing coordinates

lat
latitude column name (input data). Can be quoted or unquoted (ie. lat or 'lat').

long
longitude column name (input data). Can be quoted or unquoted (ie. long or 'long').

address
address column name (output data). Can be quoted or unquoted (ie. addr or 'addr').

return_input
if TRUE then return input dataset with geocoder results. If FALSE only the geocoder results will be included.

limit
maximum number of results to return per input coordinate. For many geocoder services the maximum value of the limit parameter is 100. Pass limit = NULL to use the default limit value of the selected geocoder service. For batch geocoding, limit must be set to 1 (default) if return_coords = TRUE To use limit > 1 or limit = NULL set return_input to FALSE. Refer to api_parameter_reference for more details.

return_coords
if TRUE return input coordinates. Defaults to TRUE if return_input is FALSE and FALSE if return_input is TRUE. This is passed to the reverse_geo() function.

unique_only
if TRUE then only unique results will be returned and return_input will be set to FALSE.

...
arguments passed to the reverse_geo function

Value

tibble (dataframe)

See also

Examples

# \donttest{
library(tibble)
library(dplyr)

reverse_geocode(
  tibble(
    latitude = c(38.895865, 43.6534817),
    longitude = c(-77.0307713,-79.3839347)
  ),
  lat = latitude,
  long = longitude,
  method = 'osm',
  full_results = TRUE,
  verbose = TRUE
)

#> Number of Unique Coordinates: 2

#> Executing single coordinate geocoding...

#> Number of Unique Coordinates: 1

#> Querying API URL: https://nominatim.openstreetmap.org/reverse

#> Passing the following parameters to the API:

#> limit : "1"

#> lat : "38.895865"

#> lon : "-77.0307713"

#> format : "json"

#> HTTP Status Code: 200

#> Query completed in: 0.2 seconds

#> Total query time (including sleep): 1 seconds

#> 

#> Number of Unique Coordinates: 1

#> Querying API URL: https://nominatim.openstreetmap.org/reverse

#> Passing the following parameters to the API:

#> limit : "1"

#> lat : "43.6534817"

#> lon : "-79.3839347"

#> format : "json"

#> HTTP Status Code: 200

#> Query completed in: 0.2 seconds

#> Total query time (including sleep): 1 seconds

#> 

#> # A tibble: 2 x 22
#>   latitude longitude address place_id licence osm_type osm_id osm_lat osm_lon
#>      <dbl>     <dbl> <chr>      <int> <chr>   <chr>     <int> <chr>   <chr>  
#> 1     38.9     -77.0 Freedo…   2.59e8 Data ©… relation 8.06e6 38.895… -77.03…
#> 2     43.7     -79.4 Toront…   1.37e8 Data ©… way      1.99e8 43.653… -79.38…
#> # … with 13 more variables: tourism <chr>, house_number <chr>, road <chr>,
#> #   quarter <chr>, city <chr>, state <chr>, postcode <chr>, country <chr>,
#> #   country_code <chr>, boundingbox <list>, amenity <chr>, neighbourhood <chr>,
#> #   state_district <chr>

louisville %>% head(5) %>% 
  reverse_geocode(lat = latitude, long = longitude, 
  method = 'arcgis', full_results = TRUE, verbose = TRUE)

#> Number of Unique Coordinates: 5

#> Executing single coordinate geocoding...

#> Number of Unique Coordinates: 1

#> Querying API URL: https://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer/reverseGeocode

#> Passing the following parameters to the API:

#> maxLocations : "1"

#> location : "-85.7976122,38.25074"

#> f : "json"

#> HTTP Status Code: 200

#> Query completed in: 0.2 seconds

#> 

#> Number of Unique Coordinates: 1

#> Querying API URL: https://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer/reverseGeocode

#> Passing the following parameters to the API:

#> maxLocations : "1"

#> location : "-85.6210708,38.2615681"

#> f : "json"

#> HTTP Status Code: 200

#> Query completed in: 0.2 seconds

#> 

#> Number of Unique Coordinates: 1

#> Querying API URL: https://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer/reverseGeocode

#> Passing the following parameters to the API:

#> maxLocations : "1"

#> location : "-85.7629019,38.2275139"

#> f : "json"

#> HTTP Status Code: 200

#> Query completed in: 1 seconds

#> 

#> Number of Unique Coordinates: 1

#> Querying API URL: https://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer/reverseGeocode

#> Passing the following parameters to the API:

#> maxLocations : "1"

#> location : "-85.6058044,38.130386"

#> f : "json"

#> HTTP Status Code: 200

#> Query completed in: 0.2 seconds

#> 

#> Number of Unique Coordinates: 1

#> Querying API URL: https://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer/reverseGeocode

#> Passing the following parameters to the API:

#> maxLocations : "1"

#> location : "-85.6539507,38.1515016"

#> f : "json"

#> HTTP Status Code: 200

#> Query completed in: 0.1 seconds

#> 

#> # A tibble: 5 x 26
#>   street city  state   zip latitude longitude address Match_addr ShortLabel
#>   <chr>  <chr> <chr> <dbl>    <dbl>     <dbl> <chr>   <chr>      <chr>     
#> 1 2722 … Loui… Kent… 40211     38.3     -85.8 2722 E… 2722 Elli… 2722 Elli…
#> 2 850 W… Loui… Kent… 40222     38.3     -85.6 242-38… 242-384 S… 242-384 S…
#> 3 1449 … Loui… Kent… 40208     38.2     -85.8 1449 S… 1449 Sain… 1449 Sain…
#> 4 9007 … Loui… Kent… 40228     38.1     -85.6 9007 S… 9007 Sage… 9007 Sage…
#> 5 376 F… Loui… Kent… 40219     38.2     -85.7 376 Fl… 376 Flirt… 376 Flirt…
#> # … with 17 more variables: Addr_type <chr>, Type <chr>, PlaceName <chr>,
#> #   AddNum <chr>, Address <chr>, Block <chr>, Sector <chr>, Neighborhood <chr>,
#> #   District <chr>, City <chr>, MetroArea <chr>, Subregion <chr>, Region <chr>,
#> #   Territory <chr>, Postal <chr>, PostalExt <chr>, CountryCode <chr>

louisville %>% head(2) %>% 
  reverse_geocode(lat = latitude, long = longitude,  
  method = 'osm', verbose = TRUE,
  limit = 5, return_input = FALSE)

#> Number of Unique Coordinates: 2

#> Executing single coordinate geocoding...

#> Number of Unique Coordinates: 1

#> Querying API URL: https://nominatim.openstreetmap.org/reverse

#> Passing the following parameters to the API:

#> limit : "5"

#> lat : "38.25074"

#> lon : "-85.7976122"

#> format : "json"

#> HTTP Status Code: 200

#> Query completed in: 0.1 seconds

#> Total query time (including sleep): 1 seconds

#> 

#> Number of Unique Coordinates: 1

#> Querying API URL: https://nominatim.openstreetmap.org/reverse

#> Passing the following parameters to the API:

#> limit : "5"

#> lat : "38.2615681"

#> lon : "-85.6210708"

#> format : "json"

#> HTTP Status Code: 200

#> Query completed in: 0.2 seconds

#> Total query time (including sleep): 1 seconds

#> 

#> # A tibble: 2 x 3
#>     lat  long address                                                           
#>   <dbl> <dbl> <chr>                                                             
#> 1  38.3 -85.8 2722, Elliott Avenue, Louisville, Jefferson County, Kentucky, 402…
#> 2  38.3 -85.6 850, Washburn Avenue, Lyndon, Jefferson County, Kentucky, 40222, …

# }